ntp: avoid jitter overflow
Compute filter jitter deltas after widening to double. Extreme but
otherwise accepted offsets could overflow int32_t during the
subtraction before the old cast took effect, which UBSan caught in the
host filter harness.

The existing 16 s jitter cap still handles huge RMS values; this only
removes the undefined signed overflow on the way to that cap.

+++ main/ntp_filter.c
@@ -43,7 +43,7 @@ void update_peer_filter(ntp_peer_t *p) {
         int n = 0;
         for (int i = 0; i < NTP_FILTER_SIZE; i++) {
             if (!p->filter[i].valid || i == best_idx) continue;
-            double d = (double)(p->filter[i].offset_us - p->best_offset_us);
+            double d = (double)p->filter[i].offset_us - (double)p->best_offset_us;
             sum_sq += d * d;
             n++;
         }
